finally a working surf

+++ src/main/java/grad/proj/recognition/train/FeatureVectorGenerator.java
@@ -3,5 +3,5 @@ package grad.proj.recognition.train;
 import grad.proj.Image;
 
 interface FeatureVectorGenerator {
-	double[] generateFeatureVector(Image image);
+	float[] generateFeatureVector(Image image);
 }
